moved pyeeg file to init file
